Move chart fixture to shared space
