+++ packages/app/src/features/farm-details/fixtures/mockChartData.ts
